create compiledriver from compile subcommand
